requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
itsdangerous==2.1.2
requests==2.31.0
markdown==3.5.1
bleach==6.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
gunicorn==21.2.0
